% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stars.R
\name{get_repo_stars}
\alias{get_repo_stars}
\alias{get_pkg_stars}
\title{Retrieve Stars of a GitHub Repo}
\usage{
get_repo_stars(repo)

get_pkg_stars(pkg)
}
\arguments{
\item{repo}{\code{character}. Repository name(s) in the form \code{user/reponame}}

\item{pkg}{\code{character}. Name of an \code{R} package.}
}
\value{
An object of class \code{c("ghstars_tbl", "data.frame")} with 2 columns:
\itemize{
\item \code{repo}: Name of the repository
\item \code{stars}: Number of stars
}
}
\description{
Retrieve the current number of stars of a GitHub repository
}
\details{
\code{get_pkg_stars()} is a shortcut for retrieving the stars of an \code{R} package.
The function tries to find the GitHub repository of the package. If it
succeeds it continues calling \code{get_repo_stars()}. If it fails either a warning
or an error is thrown depending on whether a GitHub repo couldn't be
found for some or all \code{pkg}, respectively.
}
\examples{
get_repo_stars("thomas-neitmann/mdthemes")

get_pkg_stars(c("Rcpp", "scales"))

}
\author{
Thomas Neitmann
}
