% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghstars-package.R
\docType{package}
\name{ghstars-package}
\alias{ghstars}
\alias{ghstars-package}
\title{ghstars: Retrieve GitHub Stars for any Repository}
\description{
Provides utility functions to retrieve and plot the GitHub star
    history of any repository.
}
\keyword{internal}
